+++ src/main/java/com/health/service/api/HealthServiceApiApplication.java
@@ -1,0 +1,13 @@
+package com.health.service.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HealthServiceApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HealthServiceApiApplication.class, args);
+	}
+
+}
