+++ src/main/java/com/health/service/api/common/model/APIResponseBody.java
@@ -1,0 +1,4 @@
+package com.health.service.api.common.model;
+
+public interface APIResponseBody {
+}
