+++ src/main/java/com/health/service/api/config/datasource/PrimaryDataSourceConfiguration.java
@@ -1,0 +1,15 @@
+package com.health.service.api.config.datasource;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+@Configuration
+@EnableTransactionManagement
+@EnableJpaRepositories(
+        basePackages = {
+                "com.health"
+        }
+)
+public class PrimaryDataSourceConfiguration {
+}
