+++ src/main/java/com/health/service/api/user/entity/UserEntity.java
@@ -1,0 +1,36 @@
+package com.health.service.api.user.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "User")
+public class UserEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name= "user_num")
+    private Integer userNum;
+
+    @Column(name = "user_id")
+    private String userId;
+
+    @Column(name = "password")
+    private String password;
+
+    @Column(name = "nickname")
+    private String nickName;
+
+    @Column(name = "created_on", updatable = false)
+    private LocalDateTime createdOn;
+
+    @Column(name = "modified_on")
+    private LocalDateTime modifiedOn;
+
+    @Column(name = "email")
+    private String email;
+}
