+++ src/main/java/com/health/service/api/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.health.service.api.user.repository;
+
+import com.health.service.api.user.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+}
