+++ src/main/java/com/health/service/api/user/service/UserService.java
@@ -1,0 +1,8 @@
+package com.health.service.api.user.service;
+
+import com.health.service.api.user.command.request.UserCreateRequest;
+
+public interface UserService {
+
+    Integer createUser(UserCreateRequest userCreateRequest);
+}
