--- src/main/java/com/heath/service/api/HealthServiceApiApplication.java
@@ -1,13 +1,0 @@
-package com.heath.service.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HealthServiceApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HealthServiceApiApplication.class, args);
-	}
-
-}
