--- src/main/java/com/heath/service/api/common/model/APIResponseBody.java
@@ -1,4 +1,0 @@
-package com.heath.service.api.common.model;
-
-public interface APIResponseBody {
-}
