--- src/main/java/com/heath/service/api/common/model/SingleResponse.java
@@ -1,13 +1,0 @@
-package com.heath.service.api.common.model;
-
-import lombok.Getter;
-
-@Getter
-public class SingleResponse<T> implements APIResponseBody {
-
-    private T content;
-
-    public SingleResponse(T content) {
-        this.content = content;
-    }
-}
