--- src/main/java/com/heath/service/api/user/command/request/UserCreateRequest.java
@@ -1,17 +1,0 @@
-package com.heath.service.api.user.command.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserCreateRequest {
-
-    private String userId;
-
-    private String nickName;
-
-    private String password;
-
-    private String email;
-}
