--- src/main/java/com/heath/service/api/user/entity/UserEntity.java
@@ -1,33 +1,0 @@
-package com.heath.service.api.user.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "User")
-public class UserEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "user_num")
-    private Integer userNum;
-
-    @Column(name = "user_id")
-    private String userId;
-
-    @Column(name = "nickname")
-    private String nickName;
-
-    @Column(name = "created_on", updatable = false)
-    private LocalDateTime createdOn;
-
-    @Column(name = "modified_on")
-    private LocalDateTime modifiedOn;
-
-    @Column(name = "email")
-    private String email;
-}
