--- src/main/java/com/heath/service/api/user/service/UserService.java
@@ -1,8 +1,0 @@
-package com.heath.service.api.user.service;
-
-import com.heath.service.api.user.command.request.UserCreateRequest;
-
-public interface UserService {
-
-    Integer createUser(UserCreateRequest userCreateRequest);
-}
