+++ src/main/resources/application.yaml
@@ -4,3 +4,12 @@ spring:
     url: jdbc:mysql://112.168.122.174:3306/health_service?serverTimezone=UTC&characterEncoding=UTF-8
     username: anstjaos
     password: root
+    dbcp2:
+      max-total: 10
+      initial-size: 3
+      validation-query: select 1
+  jpa:
+    show-sql: true
+    hibernate:
+      ddl-auto: none
+
