--- src/test/java/com/heath/service/api/DbUnitTestContext.java
@@ -1,19 +1,0 @@
-package com.heath.service.api;
-
-import com.github.springtestdbunit.DbUnitTestExecutionListener;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.TestExecutionListeners;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-@ActiveProfiles("h2db")
-@TestExecutionListeners({
-        DependencyInjectionTestExecutionListener.class,
-        DbUnitTestExecutionListener.class
-})
-public abstract class DbUnitTestContext {
-}
