--- src/test/java/com/heath/service/api/user/controller/UserControllerTest.java
@@ -1,50 +1,0 @@
-package com.heath.service.api.user.controller;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.heath.service.api.user.command.request.UserCreateRequest;
-import com.heath.service.api.user.service.UserService;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-@AutoConfigureMockMvc
-public class UserControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @MockBean
-    private UserService userService;
-
-    @Test
-    public void success_create_user() throws Exception {
-        UserCreateRequest request = new UserCreateRequest();
-        request.setUserId("test");
-        request.setNickName("test");
-        request.setPassword("test");
-        request.setEmail("test@asdf.com");
-
-        mockMvc.perform(MockMvcRequestBuilders.post("/users")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectToString(request)))
-                .andExpect(MockMvcResultMatchers.status().isOk());
-    }
-
-    private static String objectToString(final Object obj) {
-        try {
-            return new ObjectMapper().writeValueAsString(obj);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
